analysis/pyproject.toml
[project]
name = "hibeam-tpc-analysis"
version = "1.0.0"
description = "HIBEAM prototype TPC data analysis framework"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "matplotlib>=3.5",
    "awkward>=2.0",
    "uproot>=5.0",
    "pyyaml>=6.0",
    "mplhep>=0.3",
]

[project.optional-dependencies]
experimental = ["pandas>=1.3"]    # for CSV loading and npoints analysis
notebooks    = ["jupyterlab>=3.0"]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["hibeam*"]
